requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
fastapi>=0.68.0,<0.69.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.5
MarkupSafe==3.0.1
mysql-connector-python>=8.0.26,<8.1.0
passlib[bcrypt]>=1.7.4,<1.8.0
pyasn1==0.6.1
pydantic>=1.8.0,<2.0.0
pydantic_core==2.23.4
python-dotenv>=0.19.0,<0.20.0
python-jose[cryptography]>=3.3.0,<3.4.0
python-multipart>=0.0.5,<0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy>=1.4.0,<1.5.0
starlette
typing_extensions==4.12.2
uvicorn>=0.15.0,<0.16.0
openai>=1.0.0
aiohttp>=3.8.0
